requirements.txt

pipecat-ai[webrtc]
pipecat-ai[ollama]
fastapi
uvicorn
requests
ollama
sqlalchemy
aiohttp
python-dotenv
lightning-whisper-mlx
pyaudio
kokoro-onnx==0.4.9

-e ../pipecat